feat: introduce sequential bacnet polling with a configurable delay
This commit introduces a new feature that allows the bacnet client to
poll the BACnet network sequentially, with a configurable delay between
polls.

The delay is configurable in the configuration file of each device.

This will help mitigate the issue of the BACnet device being overwhelmed
by the number of polls.